"Refactor test plan formatting and add fullPathPlan test"
This commit centralizes the test plan formatting to a new function 'formatTestPlanToDict'. It aims to make the process of converting DataFrame to dictionary format more reusable, reducing redundant code in 'getDialogflowMessagesPlan' and 'getSignupPlan' functions.

Moreover, this commit adds the ability to test 'fullPathPlan.csv' by introducing the 'getFullPathPlan' function and updating 'getCurrentPlan' in 'automatedConversationTest.py' to use this new function. This extends the existing testing capacity.