"Added functionality to display selected drink in user input"
This change was made to ensure that the user's selected drink from the dropdown is reflected in the user input area. Previously, this functionality was missing. Consequently, the user input area didn't update with the selected drink value. Now, when the user selects the "3- Drink Choose" option, the chosen drink value is fetched and displayed in the user input area. Other files were also updated but there were no content changes, only removal of newlines at EOF.