Improve DrinkOptionComponent to handle drink quantity
Added functionality in DrinkOptionComponents to handle the quantity of drink selection. This includes addition of frames for drink type and amount and a dropdown for amounts, updating the drink message based on amount selected and handling plurals in drink names accordingly. Also, introduced gender to discriminate drinks and created an additional callback to update drink message. In OptionFrame, updated the default drink and pizza options to provide more choices to users. In DropdownComponent, fixed parent-child relationship of label and dropdown, now both will be in the same frame while earlier, label was in parent frame.