Add chained automated tests and correct beverage choice response
Introduced an automated test in chainedTests which penalises a sequence of messages through the backend and spellchecks the text responses, in an endeavor to further automate and streamline testing. The functionality of the sendEraseRequest function was modified for flexibility to allow erasure requests to be sent to different destinations. The automated test in 'automatedConversationTest' file was updated to include this change. In addition, some corrections were made to the expected output of the beverage choice response in the 'dialogflowPlan' to match the user's order accurately.